Moves Tool enum to framework folder
