+++ src/main/java/io/github/ryanlaverick/framework/file/ToolFile.java
@@ -1,9 +1,7 @@
 package io.github.ryanlaverick.framework.file;
 
 import io.github.ryanlaverick.AlchemicalTools;
-import io.github.ryanlaverick.item.Tool;
-
-import java.io.File;
+import io.github.ryanlaverick.framework.item.Tool;
 
 public final class ToolFile extends CustomFile {
     public ToolFile(Tool tool, AlchemicalTools alchemicalTools) {
